:sparkles: Update oh-my-zsh setup script URL
The URL used to download and install oh-my-zsh was updated. It previously pointed to a deprecated GitHub URL, but has now been updated to the new, proper URL for the oh-my-zsh project. Additionally, a `--unattended` flag has been added in order to allow for unattended installations during a basic setup.